requirements.txt
#SavedRetriever
praw >= 3.0.0
bleach >= 1.4.1
readability-api >= 1.0.0
imgurpython >= 1.1.6
beautifulsoup4 >= 4.4.0
-e git+https://github.com/evernote/evernote-sdk-python3#egg=evernote-1.25.2
Flask >= 0.11.1
Flask-SQLAlchemy >= 2.1
readability-lxml >= 0.6.2
Flask-WTF >= 0.12
APScheduler >= 3.2.0
sqlalchemy-migrate >= 0.10.0
waitress >= 0.9.0
flask-login >= 0.3.2
